Expand and reorganize with new plans
It's been several years since the original tech note and we've
thought a lot more about service discovery in the interim. Expand
the use cases considerably with an audit of all Phalanx applications.
Separate use cases from implementation and the current state, and
classify each into a few categories. Add an analysis of Nublado
extensions. Remove a lot of discussion of Butler, which is now
obsolete given the new Butler server.